Address review comments
* Implement promo on `DebugNoProject` view
* Move Slideshow & Promo to `packages/app`
* Consolidate Guide & Tour setup into wrapper components
* Fix styling
* Delete unused assets
