+++ packages/app/src/components/Slideshow.vue
