+++ packages/app/src/components/promo/PromoAction.cy.tsx
