+++ packages/app/src/components/promo/PromoCard.vue
