+++ packages/app/src/components/promo/PromoHeader.cy.tsx
