+++ packages/app/src/debug/empty/DebugNoProject.vue
@@ -1,26 +1,11 @@
 <template>
-  <DebugEmptyView
-    :title="t('debugPage.emptyStates.debugDirectlyInCypress')"
-    :description="t('debugPage.emptyStates.reviewRerunAndDebug')"
-    :example-test-name="t('debugPage.emptyStates.noProjectTestMessage')"
-    :help-link-sr-text="t('debugPage.emptyStates.learnAboutProjectSetupSrText')"
-    help-link-href="https://on.cypress.io/adding-new-project"
-  >
-    <template #cta="slotProps">
-      <CloudConnectButton
-        :utm-medium="DEBUG_TAB_MEDIUM"
-        :utm-content="slotProps.utmContent"
-      />
-    </template>
-  </DebugEmptyView>
+  <div class="flex h-full mx-[16px]">
+    <DebugGuide :campaign="DEBUG_PROMO_CAMPAIGNS.connectProject" />
+  </div>
 </template>
 
 <script lang="ts" setup>
-import DebugEmptyView from './DebugEmptyView.vue'
-import { useI18n } from '@cy/i18n'
-import CloudConnectButton from '../../runs/CloudConnectButton.vue'
-import { DEBUG_TAB_MEDIUM } from '../utils/constants'
-
-const { t } = useI18n()
+import DebugGuide from '../guide/DebugGuide.vue'
+import { DEBUG_PROMO_CAMPAIGNS } from '../utils/constants'
 
 </script>
