+++ packages/app/src/debug/utils/constants.ts
@@ -2,5 +2,6 @@ export const DEBUG_TAB_MEDIUM = 'Debug Tab'
 
 export const DEBUG_PROMO_CAMPAIGNS = {
   login: 'Debug Login Empty State',
+  connectProject: 'Debug Connect Project Empty State',
   recordRun: 'Debug Record Run Empty State',
 } as const
