+++ packages/app/src/runs/RunsEmpty.vue
@@ -1,72 +1,13 @@
 <template>
-  <div class="m-[10%]">
-    <Promo
+  <div class="flex h-full mx-[16px]">
+    <RunsTour
       :campaign="RUNS_PROMO_CAMPAIGNS.recordRun"
-      :medium="RUNS_TAB_MEDIUM"
-    >
-      <template #header>
-        <PromoHeader :title="t('runs.empty.title')">
-          <template #description>
-            <i18n-t
-              scope="global"
-              keypath="runs.empty.description"
-            >
-              <ExternalLink :href="guideUrl">
-                {{ t('runs.empty.link') }}
-              </ExternalLink>
-            </i18n-t>
-          </template>
-          <template #content>
-            <RecordPromptAdapter />
-          </template>
-        </PromoHeader>
-      </template>
-
-      <template #cards="{ step, goForward, reset }">
-        <TourCard
-          v-if="step === 0"
-          :action="goForward"
-        />
-        <GuideCard1
-          v-else-if="step === 1"
-          :action="goForward"
-        />
-        <GuideCard2
-          v-else-if="step === 2"
-          :action="goForward"
-        />
-        <GuideCard3
-          v-else-if="step === 3"
-          :action="reset"
-        />
-      </template>
-    </Promo>
+    />
   </div>
 </template>
 
 <script lang="ts" setup>
-import { useI18n } from '@cy/i18n'
-import { computed } from 'vue'
-import { RUNS_PROMO_CAMPAIGNS, RUNS_TAB_MEDIUM } from './utils/constants'
-import Promo from '@packages/frontend-shared/src/gql-components/promo/Promo.vue'
-import PromoHeader from '@packages/frontend-shared/src/gql-components/promo/PromoHeader.vue'
-import RecordPromptAdapter from '@cy/gql-components/RecordPromptAdapter.vue'
-import GuideCard1 from './guide/GuideCard1.vue'
-import GuideCard2 from './guide/GuideCard2.vue'
-import GuideCard3 from './guide/GuideCard3.vue'
-import TourCard from './guide/TourCard.vue'
-import ExternalLink from '@cy/gql-components/ExternalLink.vue'
-import { getUrlWithParams } from '@packages/frontend-shared/src/utils/getUrlWithParams'
+import RunsTour from './guide/RunsTour.vue'
+import { RUNS_PROMO_CAMPAIGNS } from './utils/constants'
 
-const { t } = useI18n()
-
-const guideUrl = computed(() => {
-  return getUrlWithParams({
-    url: 'https://on.cypress.io/recording-project-runs',
-    params: {
-      utm_campaign: RUNS_PROMO_CAMPAIGNS.recordRun,
-      utm_medium: RUNS_TAB_MEDIUM,
-    },
-  })
-})
 </script>
