+++ packages/app/src/runs/utils/constants.ts
@@ -2,5 +2,6 @@ export const RUNS_TAB_MEDIUM = 'Runs Tab'
 
 export const RUNS_PROMO_CAMPAIGNS = {
   login: 'Runs Login Empty State',
+  connectProject: 'Runs Connect Project Empty State',
   recordRun: 'Runs Record Run Empty State',
 } as const
