+++ packages/frontend-shared/package.json
@@ -20,7 +20,6 @@
   "devDependencies": {
     "@antfu/utils": "^0.3.0",
     "@cypress-design/css": "^0.13.1",
-    "@cypress-design/vue-icon": "0.22.1",
     "@graphql-typed-document-node/core": "^3.1.0",
     "@headlessui/vue": "1.4.0",
     "@iconify/json": "1.1.368",
